+++ lib/sh/libui.sh
@@ -65,7 +65,7 @@
 #
 #####
 
-[[ -n ${LIBUI_VERSION+x} ]] && return 0 || LIBUI_VERSION=2.002 # Sat Nov 11 15:41:32 EST 2023
+[[ -n ${LIBUI_VERSION+x} ]] && return 0 || LIBUI_VERSION=2.002 # Sun Nov 12 00:01:08 EST 2023
 
 #####
 #